% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cohesion.R
\name{calc_cohesion}
\alias{calc_cohesion}
\title{Calculate neighborhood cohesion score}
\usage{
calc_cohesion(survey_df)
}
\arguments{
\item{survey_df}{\itemize{
\item the survey dataframe including SDOH
}}
}
\value{
df_cohesion - a dataframe with person_id and cohesion score
}
\description{
Calculate neighborhood cohesion score
}
\examples{
calc_cohesion(fake_survey_df)

}
