% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_density.R
\name{calc_density}
\alias{calc_density}
\title{Calculate neighborhood density level}
\usage{
calc_density(survey_df)
}
\arguments{
\item{survey_df}{\itemize{
\item a survey dataframe including SDOH
}}
}
\value{
df_density - a dataframe with person_id and density level
}
\description{
Calculate neighborhood density level
}
\examples{
calc_density(survey_df)

}
