% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_disorder.R
\name{calc_disorder}
\alias{calc_disorder}
\title{Calculate neighborhood disorder score}
\usage{
calc_disorder(survey_df)
}
\arguments{
\item{survey_df}{\itemize{
\item a survey dataframe including SDOH
}}
}
\value{
df_disorder - a dataframe with person_id and disorder score
}
\description{
Calculate neighborhood disorder score
}
\examples{
calc_disorder(survey_df)

}
