% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spa.R
\name{calc_spa}
\alias{calc_spa}
\title{Calculate neighborhood PANES (Physical Activity Neighborhood Environment Scale) aka Support for Physical Activity (spa) score}
\usage{
calc_spa(survey_df)
}
\arguments{
\item{survey_df}{\itemize{
\item a survey dataframe including SDOH
}}
}
\value{
df_spa - a dataframe with person_id and panes score that shows support for physical activity
}
\description{
Calculate neighborhood PANES (Physical Activity Neighborhood Environment Scale) aka Support for Physical Activity (spa) score
}
\examples{
calc_spa(survey_df)

}
