% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_df.R
\docType{data}
\name{fake_survey_df}
\alias{fake_survey_df}
\title{fake_survey_df}
\format{
A data frame with three variables:
\describe{
\item{\code{person_id}}{numeric id, 1001-1010}
\item{\code{question_concept_id}}{one of the actual 81 survey question concept ids for the 81 questions in the Social Determinants of Health Survey}
\item{\code{answer}}{one of the possible answers, randomly chosen, for this question_concept_id}
\item{\code{answer_concept_id}}{one of the actual 81 survey answer concept ids for the 81 questions in the Social Determinants of Health Survey}
}

For further details on the SDOH survey for NIH AllOfUs, see \url{https://www.researchallofus.org/data-tools/survey-explorer/social-determinants-survey/}
}
\usage{
fake_survey_df
}
\description{
dataframe of fake survey data (generated in data-raw) to
enable testing of functions for survey data.
}
\keyword{datasets}
